+++ gwcelery/tests/test_tasks_first2years_external.py
@@ -1,0 +1,36 @@
+from unittest.mock import patch
+
+from . import data
+from ..tasks import first2years_external
+from ..util import read_json
+
+
+@patch('gwcelery.tasks.external_skymaps.create_upload_external_skymap')
+@patch('gwcelery.tasks.detchar.dqr_json', return_value='dqrjson')
+@patch('gwcelery.tasks.gracedb.upload.run')
+@patch('gwcelery.tasks.gracedb.get_event', return_value={
+    'graceid': 'E1', 'gpstime': 1, 'instruments': '', 'pipeline': 'Fermi',
+    'search': 'GRB',
+    'extra_attributes': {'GRB': {'trigger_duration': 1, 'trigger_id': 123,
+                                 'ra': 0., 'dec': 0., 'error_radius': 10.}},
+    'links': {'self': 'https://gracedb.ligo.org/events/E356793/'}})
+@patch('gwcelery.tasks.gracedb.create_event', return_value='E356793')
+def test_handle_create_grb_event(mock_create_event, mock_get_event,
+                                 mock_upload, mock_json,
+                                 mock_create_upload_external_skymap):
+
+    # Test IGWN alert payload.
+    alert = read_json(data, 'igwn_alert_superevent_creation.json')
+
+    alert['uid'] = 'MS180616j'
+    alert['object']['superevent_id'] = alert['uid']
+    alert['object']['preferred_event_data']['search'] = 'MDC'
+    text = first2years_external.upload_external_event(alert)
+
+    mock_create_event.assert_called_once_with(filecontents=text,
+                                              search='MDC',
+                                              pipeline='Fermi',
+                                              group='External',
+                                              labels=None)
+    mock_upload.assert_called()
+    mock_create_upload_external_skymap.assert_called()