+++ gwcelery/tests/test_tasks_raven.py
@@ -7,33 +7,37 @@ from ..tasks import gracedb
 from ..tasks import raven
 
 
+@pytest.mark.live_worker
 @pytest.mark.parametrize(
-    'group,gracedb_id,pipelines,ext_search,tl,th',
-    [['CBC', 'S1', ['Fermi', 'Swift'], 'GRB', -1, 5],
-     ['Burst', 'S2', ['Fermi', 'Swift'], 'GRB', -60, 600],
-     ['Burst', 'S3', ['SNEWS'], 'Supernova', -10, 10],
-     ['CBC', 'E1', ['Fermi'], 'GRB', -5, 1],
-     ['Burst', 'T4', ['SNEWS'], 'Supernova', -10, 10]])
-@patch('gwcelery.tasks.gracedb.create_label.run')
-@patch('gwcelery.tasks.raven.raven_pipeline.run')
-@patch('gwcelery.tasks.raven.search.run', return_value=[{'superevent_id': 'S5',
-                                                         'graceid': 'E2'}])
-@patch('gwcelery.tasks.raven.calculate_coincidence_far.run')
+    'group,gracedb_id,pipelines,ext_search,se_search,tl,th',
+    [['CBC', 'S1', ['Fermi', 'Swift'], [], ['GRB'], -1, 5],
+     ['Burst', 'S2', ['Fermi', 'Swift'], [], ['GRB'], -60, 600],
+     ['Burst', 'S3', ['SNEWS'], ['Supernova'], [], -10, 10],
+     ['Burst', 'T4', ['SNEWS'], 'Supernova', [], -10, 10],
+     ['CBC', 'MS4', ['Fermi'], ['MDC'], [], -1, 5],
+     ['CBC', 'E1', ['Fermi'], [], ['GRB'], -5, 1],
+     ['CBC', 'M1', ['Fermi'], [], ['MDC'], -5, 1]])
+@patch('gwcelery.tasks.gracedb.create_label')
+@patch('gwcelery.tasks.raven.raven_pipeline.s')
+@patch('gwcelery.tasks.raven.search.si', return_value=[{'superevent_id': 'S5',
+                                                        'graceid': 'E2'}])
+@patch('gwcelery.tasks.raven.calculate_coincidence_far')
 def test_coincidence_search(mock_calculate_coincidence_far,
                             mock_search, mock_raven_pipeline,
                             mock_create_label,
-                            group, gracedb_id, pipelines, ext_search, tl, th):
+                            group, gracedb_id, pipelines,
+                            ext_search, se_search, tl, th):
     """Test that correct time windows are used for each RAVEN search."""
     alert_object = {'superevent_id': gracedb_id}
     if 'E' in gracedb_id:
         alert_object['group'] = 'External'
     raven.coincidence_search(gracedb_id, alert_object, group,
-                             pipelines, ext_search)
+                             pipelines, ext_search, se_search)
 
     mock_search.assert_called_once_with(
-        gracedb_id, alert_object, tl, th, group, pipelines, ext_search)
+        gracedb_id, alert_object, tl, th, group, pipelines,
+        ext_search, se_search)
     mock_raven_pipeline.assert_called_once_with(
-        [{'superevent_id': 'S5', 'graceid': 'E2'}],
         gracedb_id, alert_object, tl, th, group)
 
 
@@ -50,7 +54,8 @@ def test_raven_search(mock_raven_search, event_type, event_id):
     raven.search(event_id, alert_object)
     mock_raven_search.assert_called_once_with(
         event_id, -5, 5, event_dict=alert_object,
-        gracedb=gracedb.client, group=None, pipelines=[], searches=[])
+        gracedb=gracedb.client, group=None, pipelines=[], searches=[],
+        se_searches=[])
 
 
 @pytest.mark.parametrize('group', ['CBC', 'Burst'])
@@ -126,7 +131,7 @@ def mock_get_labels(superevent_id):
 
 
 def mock_coinc_far(*args):
-    return {'temporal_coincidence_far': 1e-7,
+    return {'temporal_coinc_far': 1e-7,
             'spatiotemporal_coinc_far': None}
 
 
@@ -143,13 +148,17 @@ def mock_coinc_far(*args):
      [[], 'S13', -1, 5, 'CBC'],
      [[{'graceid': 'E4', 'pipeline': 'GRB'}], 'S14', -1, 5, 'CBC'],
      [[{'superevent_id': 'S12', 'far': .001, 'preferred_event': 'G3'}],
-        'T4', -10, 10, 'Burst']])
+        'T4', -10, 10, 'Burst'],
+     [[{'superevent_id': 'MS13', 'far': 1, 'preferred_event': 'M3'}],
+        'M15', -1, 5, 'CBC']])
 @patch('gwcelery.tasks.raven.trigger_raven_alert.run')
 @patch('gwcelery.tasks.gracedb.get_labels', mock_get_labels)
 @patch('gwcelery.tasks.raven.calculate_coincidence_far.run',
        return_value=mock_coinc_far())
+@patch('gwcelery.tasks.gracedb.update_superevent')
 @patch('gwcelery.tasks.gracedb.create_label.run')
 def test_raven_pipeline(mock_create_label,
+                        mock_update_superevent,
                         mock_calculate_coincidence_far,
                         mock_trigger_raven_alert,
                         raven_search_results, graceid, tl, th, group):
@@ -158,37 +167,57 @@ def test_raven_pipeline(mock_create_label,
     threshold, when a coincidence is found but does pass threshold, and when
     multiple events are found.
     """
-    alert_object = {'preferred_event': 'G1', 'pipeline':
-                    'Supernova' if graceid == 'T4' else 'GRB',
+    alert_object = {'preferred_event': 'G1', 'pipeline': 'Fermi',
+                    'search': 'Supernova' if graceid == 'T4' else 'GRB',
                     'labels': []}
+    time_coinc_far = mock_coinc_far()['temporal_coinc_far']
+    space_coinc_far = mock_coinc_far()['spatiotemporal_coinc_far']
     for result in raven_search_results:
         result['labels'] = []
-    if 'E' in graceid:
+    if 'S' not in graceid:
         alert_object['group'] = 'External'
-    elif 'T' in graceid:
-        alert_object['group'] = 'Test'
+        if 'T' in graceid:
+            alert_object['group'] = 'Test'
+        alert_object['graceid'] = graceid
+        for result in raven_search_results:
+            result['time_coinc_far'] = 1e-5
+            result['space_coinc_far'] = None
+    else:
+        alert_object['superevent_id'] = graceid
+        alert_object['time_coinc_far'] = 1e-5
+        alert_object['space_coinc_far'] = None
     raven.raven_pipeline(raven_search_results, graceid, alert_object, tl, th,
                          group)
 
     coinc_calls = []
     label_calls = []
+    update_calls = []
     if not raven_search_results:
         mock_calculate_coincidence_far.assert_not_called()
         mock_create_label.assert_not_called()
         return
-    elif graceid.startswith('S'):
+    if 'S' in graceid:
         for result in raven_search_results:
             label_calls.append(call('EM_COINC', result['graceid']))
             coinc_calls.append(call(alert_object, result, tl, th))
             label_calls.append(call('EM_COINC', graceid))
+            update_calls.append(
+                call(graceid, em_type=result['graceid'],
+                     time_coinc_far=time_coinc_far,
+                     space_coinc_far=space_coinc_far))
     else:
         result = raven.preferred_superevent(raven_search_results)[0]
         label_calls.append(call('EM_COINC', result['superevent_id']))
         coinc_calls.append(call(result, alert_object, tl, th))
         label_calls.append(call('EM_COINC', graceid))
+        update_calls.append(
+            call(raven_search_results[-1]['superevent_id'],
+                 em_type=graceid,
+                 time_coinc_far=time_coinc_far,
+                 space_coinc_far=space_coinc_far))
 
     alert_calls = []
-    if graceid.startswith('S'):
+    if 'S' in graceid:
         for result in raven_search_results:
             alert_calls.append(call(mock_coinc_far(),
                                     alert_object, graceid, result, group))
@@ -200,6 +229,7 @@ def test_raven_pipeline(mock_create_label,
 
     mock_calculate_coincidence_far.assert_has_calls(coinc_calls,
                                                     any_order=True)
+    mock_update_superevent.assert_has_calls(update_calls, any_order=True)
     mock_create_label.assert_has_calls(label_calls, any_order=True)
 
 
@@ -223,6 +253,37 @@ def test_preferred_superevent(raven_search_results, testnum):
     if testnum == 3:
         assert preferred_superevent == [{'superevent_id': 'S14', 'far': .0001,
                                          'preferred_event': 'G5'}]
+
+
+@pytest.mark.parametrize(
+    'new_time_far,new_space_far,old_time_far,old_space_far,result',
+    [[1e-4, None, None, None, True],
+     [1e-4, 1e-3, None, None, True],
+     [1e-4, None, 1e-5, None, False],
+     [1e-4, 1e-3, 1e-4, None, True],
+     [1e-4, 1e-3, 1e-5, 1e-6, False],
+     [1e-8, None, 1e-5, 1e-6, False],
+     [1e-4, 1e-8, 1e-4, 1e-6, True]])
+@patch('gwcelery.tasks.gracedb.update_superevent')
+def test_update_superevent(mock_update_superevent,
+                           new_time_far, new_space_far,
+                           old_time_far, old_space_far,
+                           result):
+
+    superevent = {'time_coinc_far': old_time_far,
+                  'space_coinc_far': old_space_far,
+                  'superevent_id': 'S100'}
+    ext_event = {'graceid': 'E100'}
+    coinc_far_dict = {'temporal_coinc_far': new_time_far,
+                      'spatiotemporal_coinc_far': new_space_far}
+    raven.update_coinc_far(coinc_far_dict, superevent, ext_event)
+    if result:
+        mock_update_superevent.assert_called_with(
+            'S100', em_type='E100',
+            time_coinc_far=new_time_far,
+            space_coinc_far=new_space_far)
+    else:
+        mock_update_superevent.assert_not_called()
 
 
 def _mock_get_event(graceid):
@@ -261,6 +322,11 @@ def _mock_get_event(graceid):
                 "preferred_event": "G000003",
                 "preferred_event_data": {"group": "Test"},
                 "far": 1e-6}
+    elif graceid == "MS1111":
+        return {"superevent_id": "MS1111",
+                "preferred_event": "M000004",
+                "preferred_event_data": {"group": "CBC"},
+                "far": 1e-9}
     elif graceid == 'E1':
         return {"graceid": "E1",
                 "pipeline": 'Swift',
@@ -291,8 +357,18 @@ def _mock_get_event(graceid):
                 "search": 'GRB',
                 "labels": ['NOT_GRB'],
                 "group": 'External'}
+    elif graceid == 'M5':
+        return {"graceid": "M5",
+                "pipeline": 'Fermi',
+                "search": 'MDC',
+                "labels": [],
+                "group": 'External'}
     else:
         raise AssertionError
+
+
+def _mock_get_labels(graceid):
+    return _mock_get_event(graceid)['labels']
 
 
 def _mock_get_coinc_far(graceid):
@@ -311,6 +387,9 @@ def _mock_get_coinc_far(graceid):
     elif graceid == "S9876":
         return {"temporal_coinc_far": 1e-07,
                 "spatiotemporal_coinc_far": 1e-13}
+    elif graceid == "MS1111":
+        return {"temporal_coinc_far": 1e-09,
+                "spatiotemporal_coinc_far": 1e-10}
     else:
         return {}
 
@@ -322,6 +401,7 @@ def _mock_get_coinc_far(graceid):
      ['S2468', 'E5', 'CBC', False],
      ['S2468', 'E2', 'CBC', False],
      ['S2345', 'E3', 'Burst', True],
+     ['MS1111', 'M5', 'CBC', True],
      ['E1', 'S9876', 'CBC', True],
      ['E1', 'S2468', 'CBC', False],
      ['E2', 'S5678', 'CBC', False],
@@ -329,13 +409,13 @@ def _mock_get_coinc_far(graceid):
      ['E3', 'S9876', 'Burst', True],
      ['T4', 'TS1111', 'CBC', False],
      ['E5', 'S5678', 'CBC', False]])
-@patch('gwcelery.tasks.gracedb.get_labels', return_value={})
+@patch('gwcelery.tasks.gracedb.get_labels', side_effect=_mock_get_labels)
 @patch('gwcelery.tasks.gracedb.update_superevent')
 @patch('gwcelery.tasks.gracedb.create_label.run')
 def test_trigger_raven_alert(mock_create_label, mock_update_superevent,
                              mock_get_labels,
                              graceid, result_id, group, expected_result):
-    if graceid.startswith('S'):
+    if 'S' in graceid:
         superevent_id = graceid
         ext_id = result_id
     else:
@@ -352,16 +432,6 @@ def test_trigger_raven_alert(mock_create_label, mock_update_superevent,
         label_calls = [call('RAVEN_ALERT', superevent_id),
                        call('RAVEN_ALERT', ext_id),
                        call('RAVEN_ALERT', preferred_id)]
-        if ext_event['pipeline'] == 'SNEWS':
-            time_coinc_far = space_coinc_far = None
-        else:
-            time_coinc_far = coinc_far_json['temporal_coinc_far']
-            space_coinc_far = coinc_far_json['spatiotemporal_coinc_far']
-        mock_update_superevent.assert_called_once_with(
-            superevent_id,
-            em_type=ext_id,
-            time_coinc_far=time_coinc_far,
-            space_coinc_far=space_coinc_far)
         mock_create_label.assert_has_calls(label_calls)
     else:
         mock_create_label.assert_not_called()